Server/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
mysqlclient==2.0.3
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.1
